Ivan Yesilevskyi
1. Performed migration of all Repositories to Spring Data JPA.

+++ si-core/src/main/java/com/shutafin/service/UserMatchingScoreService.java
@@ -9,6 +9,6 @@ public interface UserMatchingScoreService {
 
     UserMatchingScore getMatchingScore(User userOrigin, User userToMatch);
     Map<Long, Integer> getUserMatchingScores(User userOrigin);
-    Integer deleteUserMatchingScores(User user);
+    Long deleteUserMatchingScores(User user);
 
 }
